Extracted common options & actions in mixin
